+++ resources/js/Layouts/NavBar/Bookmark/index.jsx
