+++ resources/js/Layouts/NavBarForSp/index.jsx
