+++ resources/js/Pages/App/Company/Index/components/CompanyActionButtons/index.jsx
@@ -1,7 +1,7 @@
 import { Link } from '@inertiajs/react';
 import { AddCompanyIcon } from '@/Components/Icons/AddCompanyIcon';
 
-export default function CompanyActionButtons() {
+export function CompanyActionButtons() {
   return (
     <div className="mb-4 flex items-center justify-between">
       <Link
@@ -10,11 +10,6 @@ export default function CompanyActionButtons() {
       >
         <AddCompanyIcon />
       </Link>
-      {/* <button
-        className="rounded-full px-4 py-2 font-bold text-gray-500 transition hover:bg-gray-300 hover:text-white"
-        disabled={true}
-        dangerouslySetInnerHTML={{ __html: icons.search }}
-      ></button> */}
     </div>
   );
 }
