+++ resources/js/Pages/App/Entrysheet/Index/components/EntrysheetFilterModal/index.jsx
@@ -1,4 +1,4 @@
-export default function EntrysheetFilterModal({ showFilter, setShowFilter, filters }) {
+export function EntrysheetFilterModal({ showFilter, setShowFilter, filters }) {
   return (
     showFilter && (
       <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
