+++ resources/js/Pages/App/Entrysheet/Index/components/EntrysheetList/index.jsx
@@ -1,6 +1,6 @@
-import EntrysheetListItem from './EntrysheetListItem';
+import { EntrysheetListItem } from './EntrysheetListItem';
 
-export default function EntrysheetList({ entrysheets, onDelete }) {
+export function EntrysheetList({ entrysheets, onDelete }) {
   return (
     <div className="mt-4 grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3">
       {entrysheets.map((entrysheet) => (
